feat: refactor auth using usehooks-ts useLocalStorage as well as a custom useConfiguredAxios hook utilizing interceptors, configure node API to handle token expiry
