+++ client/package.json
@@ -7,13 +7,13 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.25.0",
-    "js-cookie": "^3.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.0",
     "sass": "^1.48.0",
     "uniqid": "^5.4.0",
+    "usehooks-ts": "^3.1.0",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
