+++ client/src/contexts/UserSettingsContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const UserSettingsContext = createContext(null);
