+++ client/src/contexts/UserSettingsSetterContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const UserSettingsSetterContext = createContext(null);
