+++ client/src/pages/ExercisesPage/ExercisesPage.jsx
@@ -1,0 +1,176 @@
+import "./ExercisesPage.scss"
+import { useState, useEffect, useCallback } from "react"
+import add from "../../assets/icons/add_black_24dp.svg"
+import AddExerciseModal from "../../components/AddExerciseModal/AddExerciseModal"
+import IndividualExercise from "../../components/IndividualExercise/IndividualExercise"
+import DeleteModal from "../../components/DeleteModal/DeleteModal"
+import EditExerciseModal from "../../components/EditExerciseModal/EditExerciseModal"
+import useConfiguredAxios from "../../hooks/useConfiguredAxios"
+const { REACT_APP_BACKEND_URL } = process.env
+
+const ExercisesPage = () => {
+  const [addExerciseModal, setAddExerciseModal] = useState(false)
+  const [deleteModal, setDeleteModal] = useState(false)
+  const [editModal, setEditModal] = useState(false)
+  const [closeModalAnimation, setCloseModalAnimation] = useState(false)
+  const [exercises, setExercises] = useState([])
+  const [currentExercise, setCurrentExercise] = useState(false)
+  const axios = useConfiguredAxios()
+
+  const getUserExercises = useCallback(() => {
+    if (axios) {
+      axios.get(`/exercises`)
+        .then(response => {
+          const sortedExercises = response.data.sort((exerciseA, exerciseB) => exerciseB.id - exerciseA.id)
+          setExercises(sortedExercises)
+        })
+        .catch(error => alert(error))
+    }
+  }, [axios])
+
+  useEffect(() => {
+    getUserExercises()
+  }, [getUserExercises])
+
+  const formatMusclesIntoString = (musclesToFormat) => {
+    return musclesToFormat.map((muscle, index) => {
+      if (index !== musclesToFormat.length - 1) {
+        return muscle + ", "
+      }
+      return muscle
+    }).join('')
+  }
+
+  const validateExerciseForm = (e, muscles, exercise) => {
+    let newExercise = {}
+    if (exercise) {
+      newExercise = { ...exercise }
+    }
+
+    newExercise.muscle = formatMusclesIntoString(muscles)
+    newExercise.name = e.target.name.value
+
+    if (!newExercise.muscle.length) {
+      alert("Please select at least one muscle before submitting a new exercise!")
+      return { error: true }
+    }
+    if (!newExercise.name || newExercise.name.length > 25) {
+      alert("Make sure to give your exercise a name less than 25 characters long!")
+      return { error: true }
+    }
+    if (newExercise.name.match(/[^A-Za-z ]/)) {
+      alert("Exercise names may not contain any numbers or special characters!")
+      return { error: true }
+    }
+    return newExercise
+  }
+
+  const addHandler = (e, muscles) => {
+    e.preventDefault()
+    const newExercise = validateExerciseForm(e, muscles)
+    if (!newExercise.error) {
+      axios.post(`/exercises`, newExercise)
+        .then(response => {
+          getUserExercises()
+          closingAnimationFunction(setAddExerciseModal)
+        })
+        .catch(error => alert(error + `\nYou may have entered a duplicate exercise name. Try using a different name than ${newExercise.name}!`))
+    }
+  }
+
+  const editExerciseHandler = (e, exercise, muscles) => {
+    e.preventDefault()
+    const newExercise = validateExerciseForm(e, muscles, exercise)
+    if (!newExercise.error) {
+      axios.put(`/exercises`, newExercise)
+        .then(response => {
+          getUserExercises()
+          closingAnimationFunction(setEditModal)
+        })
+        .catch(error => alert(error + `\nYou may have entered a duplicate exercise name. Try using a different name than ${newExercise.name}!`))
+    }
+  }
+
+  const deleteExerciseHandler = (id) => {
+    axios.delete(`/exercises/${id}`)
+      .then(response => {
+        getUserExercises()
+        closingAnimationFunction(setDeleteModal)
+      })
+      .catch(error => alert(error))
+  }
+
+  const handleSetDeleteModal = (exercise) => {
+    setCurrentExercise(exercise)
+    setDeleteModal(true)
+  }
+
+  const handleSetEditModal = (exercise) => {
+    setCurrentExercise(exercise)
+    setEditModal(true)
+  }
+
+  const closingAnimationFunction = (modalSetter) => {
+    setCloseModalAnimation(true)
+    setTimeout(() => {
+      modalSetter(false)
+      setCloseModalAnimation(false)
+    }, 300)
+  }
+
+  return (
+    <>
+      {addExerciseModal ? <AddExerciseModal
+        setAddExerciseModal={setAddExerciseModal}
+        addExerciseHandler={addHandler}
+        close={closeModalAnimation}
+      />
+        : null}
+      {editModal ? <EditExerciseModal
+        setEditExerciseModal={setEditModal}
+        editExerciseHandler={editExerciseHandler}
+        close={closeModalAnimation}
+        exercise={currentExercise}
+      />
+        : null}
+      {deleteModal ? <DeleteModal
+        setDeleteModal={setDeleteModal}
+        close={closeModalAnimation}
+        deleteHandler={deleteExerciseHandler}
+        title={currentExercise.name + " from your exercises"}
+        id={currentExercise.id}
+      />
+        : null}
+      <section className="exercises">
+        <button onClick={() => setAddExerciseModal(true)} className="exercises__add-button"><img src={add} alt="Plus sign icon" className="exercises__add" /></button>
+        <div className="exercises__top-container">
+          <h2 className="exercises__title">Exercises</h2>
+        </div>
+        <div className="exercises__scroll-container">
+          {exercises.length ?
+            <div className="exercises__bottom-container">
+              {exercises.map((exercise, index) => {
+                return <IndividualExercise
+                  key={exercise.id}
+                  exercise={exercise}
+                  index={index}
+                  setDeleteModal={exercise.user_id ? handleSetDeleteModal : null}
+                  setEditModal={exercise.user_id ? handleSetEditModal : null}
+                />
+              })}
+            </div>
+            :
+            <div className="exercises__bottom-container">
+              <div className="exercises__default-container">
+                <p className="exercises__copy">Not finding the exercise you want in our list? You can add any exercise you want to our database! Each exercise you add is only accessible to you, so it's like you're getting your own personalized exercise database!</p>
+                <button onClick={() => setAddExerciseModal(true)} className="exercises__button">Add first exercise!</button>
+              </div>
+            </div>
+          }
+        </div>
+      </section>
+    </>
+  )
+}
+
+export default ExercisesPage
