+++ server/package.json
@@ -17,7 +17,7 @@
     "epxress": "^0.0.1-security",
     "express": "^4.17.2",
     "jsonwebtoken": "^8.5.1",
-    "knex": "^1.0.1",
+    "knex": "^3.1.0",
     "mysql": "^2.18.1",
     "uniqid": "^5.4.0"
   }
