+++ server/utils.js
@@ -1,0 +1,27 @@
+require("dotenv").config();
+const { TOKEN_LIFESPAN, TOKEN_LIFESPAN_UNIT } = process.env;
+const jwt = require("jsonwebtoken");
+
+const convertToMilliseconds = (value, unit) => {
+    switch (unit) {
+        case "seconds":
+            return value * 1000;
+        case "minutes":
+            return value * 1000 * 60;
+        case "hours":
+            return value * 1000 * 60 * 60;
+        case "days":
+            return value * 1000 * 60 * 60 * 24
+        default:
+            return value;
+    }
+}
+
+const createJwt = (userId, key) => {
+    const expiration = new Date(Date.now() + convertToMilliseconds(TOKEN_LIFESPAN, TOKEN_LIFESPAN_UNIT)).toISOString();
+    return jwt.sign({ userId, expiration }, key);
+}
+
+module.exports = {
+    createJwt
+};
